requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==2.2.19
django-environ==0.4.5
flake8==3.8.4
future==0.18.2
idna==2.10
line-bot-sdk==1.19.0
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
pathspec==0.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
